requirements.txt
python-telegram-bot==20.7
yt-dlp
flask
ffmpeg-python
# python-telegram-bot
#yt-dlp
